# api/openapi.yaml
openapi: 3.0.3
...
info:
  title: FuseOne Agents API
  version: 0.1.0
  description: |
    Control API for the FuseOne Agents platform.

    This is the product's API, not a frontend backend. The Console, the CLI,
    CI pipelines and customer integrations are all first-class clients, so
    treat every change as a public contract change.

    All money is an integer in millionths of the installation's currency.
    All timestamps are RFC 3339 in UTC.
  license:
    name: Proprietary
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /runs:
    get:
      tags: [runs]
      operationId: listRuns
      summary: List runs
      parameters:
        - $ref: '#/components/parameters/Company'
        - $ref: '#/components/parameters/Area'
        - name: agentId
          in: query
          schema: { type: string }
        - name: phase
          in: query
          schema: { $ref: '#/components/schemas/Phase' }
        - name: since
          in: query
          description: Only runs started at or after this instant.
          schema: { type: string, format: date-time }
        - name: q
          in: query
          description: >
            Matches the run identifier or the agent identifier. Applied where
            the whole set is, so a page of results is the whole answer rather
            than whatever happened to be loaded.
          schema: { type: string }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of runs, newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /runs/{runId}/steps:
    get:
      tags: [runs]
      operationId: listRunSteps
      summary: Read the run's ledger trail
      description: |
        Returns the run's steps in order. This is the audit trail: every entry
        is immutable and sealed into a hash chain, so the same request always
        returns the same bytes for the same range.
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: fromSeq
          in: query
          description: First sequence number to return. Defaults to 1.
          schema: { type: integer, format: int64, minimum: 1 }
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of steps.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StepPage' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /runs/{runId}/verify:
    post:
      tags: [runs]
      operationId: verifyRun
      summary: Verify the run's hash chain
      description: |
        Walks the chain and reports whether every link is intact. Used by the
        signed export and by the tamper alarm.
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Verification result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerifyResult' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A token minted by the platform after the user authenticates against the
        installation's identity provider. For agent-initiated calls the token
        carries an actor claim, so the trail records the agent acting on behalf
        of a human rather than the human alone.
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      schema: { type: string }
    Company:
      name: company
      in: query
      description: Company scope. A single value until multi-company (PRD 3.1).
      schema: { type: string }
    Area:
      name: area
      in: query
      schema: { type: string }
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from the previous page.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    BadRequest:
      description: The request is malformed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated, but the caller does not hold the permission in this scope.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: No such resource, or it is outside the caller's scope.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem detail.
      required: [title, status]
      properties:
        type: { type: string, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
    Phase:
      type: string
      enum: [unstarted, running, awaiting_approval, awaiting_tool, parked, compensating, finished, failed]
...
    Effect:
      type: string
      description: >
        What a tool does to the world. The domain's `unknown` is deliberately
        absent: it is the zero value that makes an unclassified tool fail
        closed, never something a caller may assert.
      enum: [read, write, destructive, financial]
    StepKind:
      type: string
      enum:
        - run_started
        - planned
        - gate_decided
        - budget_reserved
        - tool_called
        - tool_returned
        - budget_reconciled
        - approval_requested
        - approval_decided
        - resumed
        - abandoned
        - compensated
        - failed
        - parked
        - run_finished
    Scope:
      type: object
      required: [company, area]
      properties:
        company: { type: string }
        area: { type: string }
    Cost:
      type: object
      description: |
        Cache tokens are reported separately from input tokens: a cache read
        costs a fraction of the price, and without the split an expensive agent
        cannot be diagnosed.
      required: [micros]
      properties:
        micros: { type: integer, format: int64 }
        inputTokens: { type: integer, format: int64 }
        outputTokens: { type: integer, format: int64 }
        cacheReadTokens: { type: integer, format: int64 }
        cacheWriteTokens: { type: integer, format: int64 }
...
    Run:
      type: object
      required: [runId, scope, agentId, versionId, phase, seq, startedAt, cost]
      properties:
        runId: { type: string }
        scope: { $ref: '#/components/schemas/Scope' }
        agentId: { type: string }
        versionId:
          type: string
          description: The specification version the run is pinned to. Publishing a new version never affects a run already in flight.
        onBehalfOf: { type: string }
        phase: { $ref: '#/components/schemas/Phase' }
        seq:
          type: integer
          format: int64
          description: Sequence number of the last recorded step.
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time, nullable: true }
        cost: { $ref: '#/components/schemas/Cost' }
        reservedMicros:
          type: integer
          format: int64
          description: Outstanding reservations not yet reconciled.
        labels:
          type: array
          description: Accumulated data-flow labels of the run context.
          items: { type: string }
        pendingApproval:
          allOf: [{ $ref: '#/components/schemas/PendingApproval' }]
          nullable: true
...
    RunPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Run' }
        nextCursor: { type: string, nullable: true }
    Step:
      type: object
      required: [seq, kind, at, hash]
      properties:
        seq: { type: integer, format: int64 }
        kind: { $ref: '#/components/schemas/StepKind' }
        at: { type: string, format: date-time }
        cost: { $ref: '#/components/schemas/Cost' }
        labels:
          type: array
          items: { type: string }
        policyHash: { type: string }
        payload:
          type: object
          additionalProperties: true
          description: |
            Kind-specific detail. Bulky content is not inlined: payloads carry
            a reference into object storage plus a digest, because arguments
            and results routinely contain personal data.
        hash:
          type: string
          description: Hex-encoded SHA-256 sealing this step to the previous one.
...
    StepPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Step' }
        nextSeq: { type: integer, format: int64, nullable: true }
...
    VerifyResult:
      type: object
      required: [valid, stepsChecked]
      properties:
        valid: { type: boolean }
        stepsChecked: { type: integer, format: int64 }
        brokenAtSeq:
          type: integer
          format: int64
          nullable: true
          description: The first step whose link failed, when valid is false.
    PendingApproval:
      type: object
      required: [runId, tool, requestedAt, atSeq]
      properties:
        runId: { type: string }
        scope: { $ref: '#/components/schemas/Scope' }
        agentId: { type: string }
        tool: { type: string }
        effect: { $ref: '#/components/schemas/Effect' }
        rule:
          type: string
          description: The rule that required approval. The trail never says "denied by policy" — it names the rule.
        reason: { type: string }
        requestedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time, nullable: true }
        atSeq: { type: integer, format: int64 }
...